fix: preserve TSM mmap across rename-aside so in-flight readers don't fault
FileStore.replace moves a still-referenced TSM file out of the way by renaming
it, but the mmap accessor's rename unmapped and remapped the file's memory.
Readers such as WalkKeys hold only a reference (not the FileStore lock) while
iterating, and the keys they return alias the mapping, so unmapping it under
them caused a use-after-free fault. The rename now leaves the mapping intact on
Unix, where a mapping survives both the rename and closing its descriptor, so
referenced readers keep working until the purger closes the file once all
references drain; Windows cannot rename a mapped file and therefore retains its
prior behavior, now documented as a limitation. This was a pre-existing bug
surfaced by the new concurrent FileStore locking test.
