+++ tsdb/engine/tsm1/reader.go
@@ -14,7 +14,6 @@ import (
 	"sync/atomic"
 
 	"github.com/influxdata/influxdb/pkg/bytesutil"
-	"github.com/influxdata/influxdb/pkg/file"
 	"github.com/influxdata/influxdb/tsdb"
 )
 
@@ -1574,49 +1573,13 @@ func (m *mmapAccessor) incAccess() {
 	atomic.AddUint64(&m.accessCount, 1)
 }
 
-func (m *mmapAccessor) rename(path string) error {
-	m.incAccess()
-
-	m.mu.Lock()
-	defer m.mu.Unlock()
-
-	err := munmap(m.b)
-	if err != nil {
-		return err
-	}
-
-	if err := m.f.Close(); err != nil {
-		return err
-	}
-
-	if err := file.RenameFile(m.f.Name(), path); err != nil {
-		return err
-	}
-
-	m.f, err = os.Open(path)
-	if err != nil {
-		return err
-	}
-
-	if _, err := m.f.Seek(0, 0); err != nil {
-		return err
-	}
-
-	stat, err := m.f.Stat()
-	if err != nil {
-		return err
-	}
-
-	m.b, err = mmap(m.f, 0, int(stat.Size()))
-	if err != nil {
-		return err
-	}
-
-	if m.mmapWillNeed {
-		return madviseWillNeed(m.b)
-	}
-	return nil
-}
+// rename moves the backing file aside, used by FileStore.replace when the file
+// is still referenced by in-flight readers. Its implementation is
+// platform-specific (see mmap_unix.go and mmap_windows.go) because in-flight
+// readers hold raw slices into the memory mapping: the index keys returned by
+// KeyAt alias m.b. On Unix the mapping survives the rename, so those readers
+// stay valid; on Windows the mapping must be released to rename, which cannot
+// preserve them. See the per-platform implementations for details.
 
 func (m *mmapAccessor) read(key []byte, timestamp int64) ([]Value, error) {
 	entry := m.index.Entry(key, timestamp)
